examples/requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
swiftpy-gcd>=0.0.1
